requirements.txt
asgiref==3.4.1
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
Pillow==8.4.0
PyJWT==2.4.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.1.1
